Device: Rewrite the datasheet and reference manual handling
The datasheet is not always the same for all the parts of a given device
file. Instead of add a field into the part definition, implement a
`datasheet_url()` method on `DeviceId` and use it.

While at it make a similar traitement for the reference manual.

Update the `DeviceInfoOut` field names to better reflect what they store.

Also rename the `tests/cargo.toml` to `cargo-toml` as `cargo test` wasn't
happy with it.